requirements.txt
cloudinary==1.44.0
Faker==37.4.0
fastapi==0.115.13
fastapi_mail==1.5.0
passlib==1.7.4
pydantic==2.11.7
pytest==8.4.1
python-dotenv==1.1.0
python_jose==3.3.0
redis==6.2.0
slowapi==0.1.9
SQLAlchemy==2.0.41
starlette==0.47.1
pytest-cov==6.2.1
sphinx==7.4.7